+++ caas/kubernetes/provider/storage.go
@@ -11,11 +11,8 @@ import (
 	"github.com/juju/errors"
 	"github.com/juju/schema"
 	core "k8s.io/api/core/v1"
-	storagev1 "k8s.io/api/storage/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	k8slabels "k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/types"
 
 	"github.com/juju/juju/environs/context"
 	"github.com/juju/juju/storage"
@@ -71,91 +68,6 @@ func validateStorageAttributes(attributes map[string]interface{}) error {
 		return errors.Trace(err)
 	}
 	return nil
-}
-
-// StorageProviderTypes is defined on the storage.ProviderRegistry interface.
-func (k *kubernetesClient) StorageProviderTypes() ([]storage.ProviderType, error) {
-	return []storage.ProviderType{K8s_ProviderType}, nil
-}
-
-// StorageProvider is defined on the storage.ProviderRegistry interface.
-func (k *kubernetesClient) StorageProvider(t storage.ProviderType) (storage.Provider, error) {
-	if t == K8s_ProviderType {
-		return &storageProvider{k}, nil
-	}
-	return nil, errors.NotFoundf("storage provider %q", t)
-}
-
-func (k *kubernetesClient) deleteStorageClasses(selector k8slabels.Selector) error {
-	err := k.client().StorageV1().StorageClasses().DeleteCollection(&v1.DeleteOptions{
-		PropagationPolicy: &defaultPropagationPolicy,
-	}, v1.ListOptions{
-		LabelSelector: selector.String(),
-	})
-	if k8serrors.IsNotFound(err) {
-		return nil
-	}
-	return errors.Annotate(err, "deleting model storage classes")
-}
-
-func (k *kubernetesClient) listStorageClasses(selector k8slabels.Selector) ([]storagev1.StorageClass, error) {
-	listOps := v1.ListOptions{
-		LabelSelector: selector.String(),
-	}
-	list, err := k.client().StorageV1().StorageClasses().List(listOps)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	if len(list.Items) == 0 {
-		return nil, errors.NotFoundf("storage classes with selector %q", selector)
-	}
-	return list.Items, nil
-}
-
-func (k *kubernetesClient) ensurePVC(pvc *core.PersistentVolumeClaim) (*core.PersistentVolumeClaim, func(), error) {
-	cleanUp := func() {}
-	out, err := k.createPVC(pvc)
-	if err == nil {
-		// Only do cleanup for the first time!
-		cleanUp = func() { _ = k.deletePVC(out.GetName(), out.GetUID()) }
-		return out, cleanUp, nil
-	}
-	if !errors.IsAlreadyExists(err) {
-		return nil, cleanUp, errors.Trace(err)
-	}
-	existing, err := k.getPVC(pvc.GetName())
-	if err != nil {
-		return nil, cleanUp, errors.Trace(err)
-	}
-	// PVC is immutable after creation except resources.requests for bound claims.
-	// TODO(caas): support requests - currently we only support limits which means updating here is a no ops for now.
-	existing.Spec.Resources.Requests = pvc.Spec.Resources.Requests
-	out, err = k.updatePVC(existing)
-	return out, cleanUp, errors.Trace(err)
-}
-
-func (k *kubernetesClient) createPVC(pvc *core.PersistentVolumeClaim) (*core.PersistentVolumeClaim, error) {
-	out, err := k.client().CoreV1().PersistentVolumeClaims(k.namespace).Create(pvc)
-	if k8serrors.IsAlreadyExists(err) {
-		return nil, errors.AlreadyExistsf("PVC %q", pvc.GetName())
-	}
-	return out, errors.Trace(err)
-}
-
-func (k *kubernetesClient) updatePVC(pvc *core.PersistentVolumeClaim) (*core.PersistentVolumeClaim, error) {
-	out, err := k.client().CoreV1().PersistentVolumeClaims(k.namespace).Update(pvc)
-	if k8serrors.IsNotFound(err) {
-		return nil, errors.NotFoundf("PVC %q", pvc.GetName())
-	}
-	return out, errors.Trace(err)
-}
-
-func (k *kubernetesClient) deletePVC(name string, uid types.UID) error {
-	err := k.client().CoreV1().PersistentVolumeClaims(k.namespace).Delete(name, newPreconditionDeleteOptions(uid))
-	if k8serrors.IsNotFound(err) {
-		return nil
-	}
-	return errors.Trace(err)
 }
 
 type storageProvider struct {
@@ -249,7 +161,7 @@ func getStorageMode(attrs map[string]interface{}) (*core.PersistentVolumeAccessM
 		case "ReadWriteOnce", "RWO":
 			out = core.ReadWriteOnce
 		default:
-			return &out, errors.NotSupportedf("storage mode %q", m)
+			return nil, errors.NotSupportedf("storage mode %q", m)
 		}
 		return &out, nil
 	}
